fix(ado-script): exclude smoke test from default vitest run; consolidate shared tests
- Add vitest.config.ts with `include: ["src/**/*.test.ts"]` so the
  default `npm test` no longer picks up test/smoke.test.ts (which
  requires the ncc bundle to exist and was failing in CI before the
  build step ran).
- Add vitest.config.smoke.ts targeting test/ for the smoke run.
- Wire `npm run test:smoke` and the CI Smoke-test step to use the new
  smoke config (`vitest run -c vitest.config.smoke.ts`).
- Move shared module tests under src/shared/__tests__/ to mirror the
  layout used by src/gate/__tests__/. Update relative imports
  (./foo.js → ../foo.js) and the vi.mock path in ado-client.test.ts.

Validation:
- `npm test` → 171/171 ✅ (smoke excluded)
- `npx vitest run -c vitest.config.smoke.ts` after `npm run build`
  → 2/2 ✅
- `npm run typecheck` ✅
